Move from assert() to own verify()
Depending on the version of PHP and the settings in php.ini, assert()
might be a disabled function that does not do anything.

The way we use assert in the web-framework is to make sure that code
paths that should never happen are blocked. Silently having them
disabled opens web-framework up for possible security issues.

Moving to verify() with the same functionality as assert() fixes this,
although it does require all assert() statements to be rewritten in the
new verify() form.