Rewrote logic around email_is_username and login_with_email
We now use unique_identifier, and mandate that the username field is
always either a unique 'name' or unique 'email' address..

Username is used for all actions and in case email is the unique
identifier, the username should also be changed when changing a user's
email address.

Username is always used as the unique identifier row in the table.