Required App DB version should be a minimum.
In situations where there is a dev / staging / production situation
where more than one stage share a single database, the required app DB
version indicated in the config should reflect the minimum App DB
version that the app is looking for.

This introduces a risk of incompatible database changes in the future,
not properly handled. When we support those changes, we should introduce
a kind of 'forward-compatibility-version' or similar mechanic.